# test(broker): Optimize integration tests