Optimized some tests by removing waits:
- MQTT `Bridge`: 34s -> 11s
- `InspectRandomNodeHelper`: ~64s -> ~58s (maybe waits are not needed when fast-chain is used)